+++ TMDT/TMDT/Models/SanPhamNhapKho.cs
@@ -14,10 +14,10 @@ namespace TMDT.Models
     
     public partial class SanPhamNhapKho
     {
-        public int MaSanPhamNhapKho { get; set; }
-        public int MaSanPham { get; set; }
+        public long MaSanPhamNhapKho { get; set; }
+        public long MaSanPham { get; set; }
         public Nullable<int> SoLuong { get; set; }
-        public Nullable<int> GiaNhap { get; set; }
+        public Nullable<double> GiaNhap { get; set; }
         public Nullable<System.DateTime> NgayTao { get; set; }
     
         public virtual SanPham SanPham { get; set; }
